requirements.txt
selenium>=4.0.0
pillow>=8.0.0
webdriver-manager>=3.5.0